+++ src/client/ui/demos.c
@@ -48,7 +48,7 @@ DEMOS MENU
 
 typedef struct {
     unsigned    type;
-    size_t      size;
+    int64_t     size;
     time_t      mtime;
     char        name[1];
 } demoEntry_t;
@@ -58,12 +58,12 @@ typedef struct m_demos_s {
     menuList_t      list;
     int             numDirs;
     uint8_t         hash[16];
-    char    browse[MAX_OSPATH];
-    int     selection;
-    int     year;
-    int     widest_map, widest_pov;
-    size_t  total_bytes;
-    char    status[32];
+    char            browse[MAX_OSPATH];
+    int             selection;
+    int             year;
+    int             widest_map, widest_pov;
+    uint64_t        total_bytes;
+    char            status[32];
 } m_demos_t;
 
 static m_demos_t    m_demos;
